+++ src/App.js
@@ -9,7 +9,7 @@ import ReactGA from 'react-ga';
 
 // Pages
 import Homepage from './pages/homepage'
-import Editor from './pages/editor'
+import Editor from './pages/editors'
 
 ReactGA.initialize("G-VTPVN058D5");
 
@@ -22,7 +22,7 @@ function App() {
       <Router>
         <Switch>
           <Route component={Homepage} exact path="/" />
-          <Route component={Editor} path="/invoice-edit" />
+          <Route component={Editor} path="/edit" />
         </Switch>
       </Router>
     </ChakraProvider>